cloudrun/requirements.txt
firebase-admin==4.2.0
flask==1.1.2
google-cloud-storage==1.28.1
gunicorn==20.0.4
numpy==1.18.4
joblib==0.14.1
pillow==7.1.2
requests==2.23.0
